fix path issues, global hydra fixed
