+++ requirements_inference.txt
@@ -7,3 +7,4 @@ fastapi>=0.68.0,<0.69.0
 uvicorn>=0.15.0,<0.16.0
 pandas
 hydra-core
+zipfile36
