+++ src/configs/config.yaml
@@ -20,4 +20,4 @@ paths:
   model_path: 'models/model.pth'
   model_config_path: 'src/configs/model_configs/hyperparams.json'
 tuning:
-  num_configs: 200
+  num_configs: 250
